requirements.txt
pytest==9.0.2
selenium==4.41.0
requests==2.32.5
allure-pytest==2.15.3